Only one piece of code generating User-Agent HTTP headers
